% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{minute2time}
\alias{minute2time}
\title{minute2time}
\usage{
minute2time(x)
}
\arguments{
\item{x}{minutes ex: 120}
}
\value{
time of day
}
\description{
convert minutes to time of day
}
\examples{
minute2time(120)
2:00
}
\author{
Alok Kumar <connect.alokkumar@gmail.com>
}
