% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time2min}
\alias{time2min}
\title{time2min}
\usage{
time2min(x)
}
\arguments{
\item{x}{minutes ex: 1122}
}
\value{
time of day
}
\description{
convert time of day (format hhmm) to minutes
}
\examples{
time2min(1200)
720
}
\author{
Alok Kumar <connect.alokkumar@gmail.com>
}
