% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timser.R
\docType{package}
\name{timser}
\alias{timser}
\alias{timser-package}
\title{timser: This wonderul package is an ideal tool-box for working with time series.}
\description{
timser: This wonderul package is an ideal tool-box for working with time series.
}
\section{Class TIME.SERIES}{

timser provides a Reference class named as TIME.SERIES.
This package supports the latest forecasting techniques for prediction.
It also supports the latest visualization techniques for presentation.
}

