% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdaily.R
\docType{class}
\name{TS.DAILY-class}
\alias{TS.DAILY-class}
\alias{TS.DAILY}
\title{A Reference Class representing a time series.}
\description{
A Reference Class representing a time series.
}
\section{Fields}{

\describe{
\item{\code{N.int}}{An integer representing the number of time intervals in the time series}

\item{\code{ctn}}{An integer representing the current time interval number}

\item{\code{stn}}{An integer representing the starting time interval number of the control window}

\item{\code{etn}}{An integer representing the starting time interval number of the control window}

\item{\code{data}}{A data.frame with the same number of rows as fields \code{time} containing the time series data}
}}

\section{Methods}{

\describe{
\item{\code{feedEventLog.molten(dataset, time_col, var_col, value_col,
  aggrFunctions = NULL)}}{                            feeds additional data to the time series in molten format.

                            }

\item{\code{sortByDate(decreasing = F)}}{Sorts the data based on date}
}}

